update user-slice
move actions to user-slice-action.js
add 'activate account' and 'signin' thunks
