+++ src/features/user/user-slice.js
@@ -1,37 +1,8 @@
-// https://github.com/christofferbergj/react-redux-toolkit-example/blob/master/src/features/users/usersSlice.ts
-import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { calculateOffset } from './calculate-offset';
+import { createSlice } from '@reduxjs/toolkit';
 import { postCard } from '../freedictionary/freedictionary-slice';
-import makeErrorSerializable from 'helpers/make-error-serializable';
+import { fetchUser, signIn } from './user-slice-actions';
 
 const token = JSON.parse(localStorage.getItem('token'));
-
-export const signUp = createAsyncThunk(
-  '@@user/sign-up',
-  async ({ firstName, email, password }, { extra, rejectWithValue }) => {
-    try {
-      const result = await extra.authApi.signUp({
-        firstName,
-        email,
-        password,
-      });
-      return result;
-    } catch (error) {
-      return rejectWithValue(makeErrorSerializable(error));
-    }
-  }
-);
-
-export const fetchUser = createAsyncThunk(
-  '@@user/fetch-user',
-  async (tgid, { extra }) => {
-    const result = await extra.jsonServerApi.fetchUserAttributes({
-      tgid,
-      timezone: calculateOffset(),
-    });
-    return result;
-  }
-);
 
 const initialState = {
   user: null,
@@ -49,6 +20,11 @@ const userSlice = createSlice({
     },
     [postCard.fulfilled]: (state) => {
       if (state.user?.totalCards) state.user.totalCards += 1;
+    },
+    [signIn.fulfilled]: (state, action) => {
+      state.token = action.payload.token;
+      state.user = action.payload.user;
+      localStorage.setItem('token', JSON.stringify(action.payload.token));
     },
   },
 });
